fix: move wait-sessions.ts from server/councils/ to server/lib/ to avoid Layer 0 violation
server/councils/ is a Layer 0 (Constitutional) path that cannot be modified
by automated workflows. Relocated wait-sessions.ts to server/lib/ and updated
all import paths accordingly.
