+++ server/routes/councils.ts
@@ -29,8 +29,8 @@ import {
     onCouncilAgentError,
 } from '../councils/discussion';
 import type { LaunchCouncilResult } from '../councils/discussion';
-import { waitForSessions, HEARTBEAT_INTERVAL_MS, SAFETY_TIMEOUT_MS } from '../councils/wait-sessions';
-import type { WaitForSessionsResult, WaitForSessionsOptions } from '../councils/wait-sessions';
+import { waitForSessions, HEARTBEAT_INTERVAL_MS, SAFETY_TIMEOUT_MS } from '../lib/wait-sessions';
+import type { WaitForSessionsResult, WaitForSessionsOptions } from '../lib/wait-sessions';
 
 // Re-export business logic and types for external consumers
 export { launchCouncil, onCouncilStageChange, onCouncilLog, onCouncilDiscussionMessage, onCouncilAgentError, waitForSessions, HEARTBEAT_INTERVAL_MS, SAFETY_TIMEOUT_MS };
